PT_web/requirements.txt
flask
cryptography
mysql-connector-python
